xdim and ydim are now size_t
Former-commit-id: 875a7aeb5e55d66468a2cdf7f42709f5ce6020f7

+++ applications/programs/classify_kerdensom/classify_kerdensom_main.cpp
@@ -39,8 +39,8 @@ public:
     double         eps;          // Stopping criteria
     unsigned       iter;         // Iteration number
     bool           norm;         // Normalize?
-    int            xdim;         // X-dimension (-->)
-    int            ydim;         // Y-dimension
+    size_t         xdim;         // X-dimension (-->)
+    size_t         ydim;         // Y-dimension
     double         reg0;         // Initial reg
     double         reg1;         // Final reg
     std::string    layout;       // layout (Topology)
